Extending the Kubernetes event logging to cover Pod events (#91)
* Extending the Kubernetes event logging to cover Pod events

In #87/#88, we logged additional information from the Kubernetes Event API when
a Job couldn't create a Pod.  In this change, we expand that to log additional
event information when a Pod can't start running.  This covers cases like
`ErrImagePull` or pod scheduling constraint failures that will prevent the Pod
from going into a Running state.

Closes #90

* Python 3.8 type hint compatibility